Unit tests fix
